# Conflicts:
#	src/test/java/org/rulelearn/types/ComparableExtTest.java

+++ src/main/java/org/rulelearn/types/IntegerFieldFactory.java
@@ -16,6 +16,7 @@
 
 package org.rulelearn.types;
 
+import org.rulelearn.core.TernaryLogicValue;
 import org.rulelearn.data.AttributePreferenceType;
 
 /**
@@ -72,7 +73,8 @@ public class IntegerFieldFactory {
 	 * @return created field
 	 */
 	public IntegerField clone (IntegerField field) {
-		return field.selfClone();
+		// TODO check if correct
+		return (IntegerField)field.selfClone();
 	}
 	
 	/**
